third_party/blink/web_tests/fast/css-generated-content/pseudo-animation-before-onload-expected.txt

Animations on :before and :after pseudo elements should run when applied before onload.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS div.offsetWidth is within 1 of 20
PASS div.offsetWidth is within 1 of 20
PASS div.offsetWidth is within 1 of 12
PASS div.offsetWidth is within 1 of 12
     
